ml-service/requirements.txt
# ML Service Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML Libraries
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.0

# NLP
nltk==3.8.1

# Database
pymongo==4.6.0
motor==3.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
